fix: delete comments

+++ test/data/model.js
@@ -22,7 +22,8 @@ describe("Model", function() {
 describe("ModelStore", function() {
     this.timeout(30000);
     beforeEach(async function() {
-        const uri = "mongodb://localhost/enas";
+        const uri = await yonius.confP("MONGO_URL");
+        if (!uri) return this.skip();
         await yonius.initMongo(mongoose, uri);
         await mongoose.connection.db.dropDatabase();
     });
